size_t resolve

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: gneve <gneve@student.s19.be>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/17 01:38:47 by gneve             #+#    #+#             */
-/*   Updated: 2020/11/24 10:23:08 by gneve            ###   ########.fr       */
+/*   Updated: 2020/11/27 06:13:23 by gneve            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int i;
